Ajuste no type de rating, implementação da camada de controllers

+++ service_feedback/src/main/java/com/sistemareserva/service_feedback/client/broker/Producer.java
@@ -1,5 +1,7 @@
 package com.sistemareserva.service_feedback.client.broker;
 
+import java.math.BigDecimal;
+
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.stereotype.Service;
 
@@ -14,7 +16,7 @@ public class Producer {
     private final RabbitTemplate rabbitTemplate;
     private final String exchange = "rating-exchange";
 
-    public void sendUpdateRating(Long idQuarto, Double rating) {
+    public void sendUpdateRating(Long idQuarto, BigDecimal  rating) {
         UpdateRatingRequest updateRatingRequest = new UpdateRatingRequest(idQuarto, rating);
         rabbitTemplate.convertAndSend(exchange, null, updateRatingRequest);
     }
